Add support (untested) for EP2316 devices.
The EP23[012]x devices have distinct input and output ports, unlike the
EP23[34]x devices, which can use any port for input or output.

This *should* Just Work, as long as the correct number of PDOs are
defined for the device.  The difference is the "input3" and "output3" on
an EP2338 share a port, while on an EP2316 they're two individual
connectors.