Create c-cpp.yml
Create an action to do test builds, to make sure that merges, etc, still build.

This is somewhat tricky because LinuxCNC really only supports Debian, and Github mostly only provides Ubuntu.  We'll try using Ubuntu's `linuxcnc-uspace` package; it might work.

This is probably going to fail due to a missing Ethercat package, but we can deal with that next.