build: Move types[] from lcec_main into individual drivers
Beginning of removing the global types[] list in `lcec_main.c` and
moving the code into individual drivers.  This will make patching less
complicated as it'll be one fewer file to edit when adding new drivers.

The goal is to be able to simply drop a pair of `lcec_FOO.[ch]` files
into `src/` and have them compile and work.  We're not there quite yet.

Issue: #35