feat(el3xxx): Add initial support for EL3204
This adds support for the EL3204 4-channel temperature sensor to the
el3xxx driver.  It also moves my test setup from `examples/testdata` to
`tests`.

This is still in flux and has too many debugging statements in it, but
mostly works.  The device shows up and seems to function with a Pt100 on
one channel and a Pt1000 on another.  One claims that it's 17.4C and the
other claims it's 20.1C.  I think that's about as good as I can expect
for now.

This adds some parameters to the el3xxx driver for the first time; the
EL3204 (and *only* the EL3204 for now) takes parameters like this:

```
<slave idx="EL3204" name="D16">
  <modParam name="sensor0" value="Pt100"/>
  <modParam name="sensor2" value="Pt1000"/>
</slave>
```

The exact syntax will probably change slightly in the near future.  For
now, the `value` is case-insensitive and needs to be one of:

- Pt100
- Pt200
- Pt500
- Pt1000
- Ni100
- Ni120
- Ni1000
- Ni1000-TK5000
- ohm/16
- ohm/64

The two "ohm" settings disable most of the temperature compensation
logic and return the measured Ohms.  `ohm/16` returns the measured
resistance in units of 1/16th of an Ohm, and `ohm/64` returns the
measured resistance in units of 1/64th of an Ohm.

There are a few other sensor types defined in
https://download.beckhoff.com/download/Document/io/ethercat-terminals/el32xxen.pdf#page=223,
but they seem to need additional calibration and are probably out of
scope here.